Update README.md
The password strength project aims to develop a password-strength testing tool. It will provide users with an easy and efficient way to evaluate the strength of their passwords. The tool will analyze various factors such as length, complexity, and inclusion of special characters to determine the strength level. Additionally, it will provide suggestions and tips for creating stronger passwords. The project will focus on creating a user-friendly interface as a web/desktop application and incorporating robust algorithms to assess password strength accurately. Ultimately, the goal is to enhance cybersecurity awareness and empower users to protect their accounts with strong passwords.